+++ src/WinRT.Runtime/Projections/NotifyCollectionChangedEventHandler.cs
@@ -164,4 +164,38 @@ namespace ABI.System.Collections.Specialized
             return 0;
         }
     }
+
+    internal sealed unsafe class NotifyCollectionChangedEventHandlerEventSource : EventSource<global::System.Collections.Specialized.NotifyCollectionChangedEventHandler>
+    {
+        internal NotifyCollectionChangedEventHandlerEventSource(
+            IObjectReference objectReference,
+#if NET
+            delegate* unmanaged[Stdcall]<IntPtr, IntPtr, EventRegistrationToken*, int> addHandler,
+#else
+            delegate* unmanaged[Stdcall]<IntPtr, IntPtr, out EventRegistrationToken, int> addHandler,
+#endif
+            delegate* unmanaged[Stdcall]<IntPtr, EventRegistrationToken, int> removeHandler)
+            : base(objectReference, addHandler, removeHandler)
+        {
+        }
+
+        protected override ObjectReferenceValue CreateMarshaler(global::System.Collections.Specialized.NotifyCollectionChangedEventHandler del) =>
+            NotifyCollectionChangedEventHandler.CreateMarshaler2(del);
+
+        protected override EventSourceState<global::System.Collections.Specialized.NotifyCollectionChangedEventHandler> CreateEventSourceState() =>
+            new EventState(ObjectReference.ThisPtr, Index);
+
+        private sealed class EventState : EventSourceState<global::System.Collections.Specialized.NotifyCollectionChangedEventHandler>
+        {
+            public EventState(IntPtr obj, int index)
+                : base(obj, index)
+            {
+            }
+
+            protected override global::System.Collections.Specialized.NotifyCollectionChangedEventHandler GetEventInvoke()
+            {
+                return (obj, e) => targetDelegate?.Invoke(obj, e);
+            }
+        }
+    }
 }
